Don't mix up display logic with data
